specialize to avoid missing pgid error on windows
